+++ lib/views/pc/widgets/scatter.dart
@@ -161,7 +161,8 @@ class ScatterData {
   final double yBallPointsStddev;
 }
 
-const String query = """
+const String query =
+    """
 query MyQuery {
   team {
     number
@@ -220,16 +221,16 @@ Future<List<ScatterData>> fetchScatterData() async {
                 final double xBallPointsAvg =
                     avgTeleLower + avgAutoLower + avgTeleUpper + avgAutoUpper;
                 final List<dynamic> matches = e["matches"] as List<dynamic>;
-                final List<double> matchBallPoints = matches
+                final List<int> matchBallPoints = matches
                     .map(
-                      (final dynamic e) => ((e["auto_lower"] as double) * 2 +
-                          (e["tele_lower"] as double) * 1 +
-                          (e["auto_upper"] as double) * 4 +
-                          (e["tele_upper"] as double) * 2),
+                      (final dynamic e) => ((e["auto_lower"] as int) * 2 +
+                          (e["tele_lower"] as int) * 1 +
+                          (e["auto_upper"] as int) * 4 +
+                          (e["tele_upper"] as int) * 2),
                     )
                     .toList();
                 double yStddevBallPoints = 0;
-                for (final double element in matchBallPoints) {
+                for (final int element in matchBallPoints) {
                   yStddevBallPoints += (element - xBallPointsAvg).abs();
                 }
                 yStddevBallPoints /= matchBallPoints.length;
